[refactor] Move utils to bitbuilder/internal
