+++ src/main/kotlin/io/vexelabs/bitbuilder/internal/Extensions.kt
@@ -1,4 +1,4 @@
-package io.vexelabs.bitbuilder.llvm.internal.util
+package io.vexelabs.bitbuilder.internal
 
 import org.bytedeco.javacpp.Pointer
 import org.bytedeco.javacpp.PointerPointer
@@ -24,3 +24,19 @@ internal inline fun <reified P : Pointer, R> PointerPointer<P>.map(
 
     return res
 }
+
+/**
+ * Util function to convert kotlin [Boolean] to [Int]
+ *
+ * Used because LLVM C api does not use booleans, it uses
+ * C integers 1 and 0.
+ */
+internal fun Boolean.toLLVMBool() = if (this) 1 else 0
+
+/**
+ * Util function to convert kotlin [Int] to [Boolean]
+ *
+ * Used because LLVM C api does not use booleans, it uses
+ * C integers 1 and 0.
+ */
+internal fun Int.fromLLVMBool() = this > 0
