+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/executionengine/GenericValue.kt
@@ -2,7 +2,7 @@ package io.vexelabs.bitbuilder.llvm.executionengine
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
 import io.vexelabs.bitbuilder.llvm.internal.contracts.Disposable
-import io.vexelabs.bitbuilder.llvm.internal.util.toLLVMBool
+import io.vexelabs.bitbuilder.internal.toLLVMBool
 import io.vexelabs.bitbuilder.llvm.ir.types.FloatType
 import io.vexelabs.bitbuilder.llvm.ir.types.IntType
 import org.bytedeco.javacpp.Pointer
