+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/executionengine/callbacks/MemoryManagerAllocateDataSectionCallback.kt
@@ -1,7 +1,7 @@
 package io.vexelabs.bitbuilder.llvm.executionengine.callbacks
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.Callback
-import io.vexelabs.bitbuilder.llvm.internal.util.fromLLVMBool
+import io.vexelabs.bitbuilder.internal.fromLLVMBool
 import org.bytedeco.javacpp.BytePointer
 import org.bytedeco.javacpp.Pointer
 import org.bytedeco.llvm.LLVM.LLVMMemoryManagerAllocateDataSectionCallback
