+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/Context.kt
@@ -2,8 +2,8 @@ package io.vexelabs.bitbuilder.llvm.ir
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
 import io.vexelabs.bitbuilder.llvm.internal.contracts.Disposable
-import io.vexelabs.bitbuilder.llvm.internal.util.fromLLVMBool
-import io.vexelabs.bitbuilder.llvm.internal.util.toLLVMBool
+import io.vexelabs.bitbuilder.internal.fromLLVMBool
+import io.vexelabs.bitbuilder.internal.toLLVMBool
 import io.vexelabs.bitbuilder.llvm.ir.callbacks.DiagnosticHandlerBase
 import io.vexelabs.bitbuilder.llvm.ir.callbacks.DiagnosticHandlerCallback
 import io.vexelabs.bitbuilder.llvm.ir.callbacks.YieldCallback
