+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/attributes/Attribute.kt
@@ -1,7 +1,7 @@
 package io.vexelabs.bitbuilder.llvm.ir.attributes
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
-import io.vexelabs.bitbuilder.llvm.internal.util.fromLLVMBool
+import io.vexelabs.bitbuilder.internal.fromLLVMBool
 import org.bytedeco.llvm.LLVM.LLVMAttributeRef
 import org.bytedeco.llvm.global.LLVM
