+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/instructions/traits/CallBase.kt
@@ -1,7 +1,7 @@
 package io.vexelabs.bitbuilder.llvm.ir.instructions.traits
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
-import io.vexelabs.bitbuilder.llvm.internal.util.map
+import io.vexelabs.bitbuilder.internal.map
 import io.vexelabs.bitbuilder.llvm.ir.AttributeIndex
 import io.vexelabs.bitbuilder.llvm.ir.CallConvention
 import io.vexelabs.bitbuilder.llvm.ir.Type
