+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/instructions/traits/MemoryAccessor.kt
@@ -1,8 +1,8 @@
 package io.vexelabs.bitbuilder.llvm.ir.instructions.traits
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
-import io.vexelabs.bitbuilder.llvm.internal.util.fromLLVMBool
-import io.vexelabs.bitbuilder.llvm.internal.util.toLLVMBool
+import io.vexelabs.bitbuilder.internal.fromLLVMBool
+import io.vexelabs.bitbuilder.internal.toLLVMBool
 import io.vexelabs.bitbuilder.llvm.ir.AtomicOrdering
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
