+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/values/constants/ConstantFloat.kt
@@ -1,12 +1,12 @@
 package io.vexelabs.bitbuilder.llvm.ir.values.constants
 
-import io.vexelabs.bitbuilder.llvm.internal.util.fromLLVMBool
+import io.vexelabs.bitbuilder.internal.fromLLVMBool
 import io.vexelabs.bitbuilder.llvm.ir.RealPredicate
 import io.vexelabs.bitbuilder.llvm.ir.types.FloatType
 import io.vexelabs.bitbuilder.llvm.ir.types.IntType
 import io.vexelabs.bitbuilder.llvm.ir.values.ConstantValue
-import io.vexelabs.bitbuilder.raii.resourceScope
-import io.vexelabs.bitbuilder.raii.toResource
+import io.vexelabs.bitbuilder.internal.resourceScope
+import io.vexelabs.bitbuilder.internal.toResource
 import org.bytedeco.javacpp.IntPointer
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
