+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/values/constants/ConstantVector.kt
@@ -1,7 +1,7 @@
 package io.vexelabs.bitbuilder.llvm.ir.values.constants
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.Unreachable
-import io.vexelabs.bitbuilder.llvm.internal.util.toLLVMBool
+import io.vexelabs.bitbuilder.internal.toLLVMBool
 import io.vexelabs.bitbuilder.llvm.ir.IntPredicate
 import io.vexelabs.bitbuilder.llvm.ir.Value
 import io.vexelabs.bitbuilder.llvm.ir.types.FloatType
