+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/values/traits/DebugLocationValue.kt
@@ -1,8 +1,8 @@
 package io.vexelabs.bitbuilder.llvm.ir.values.traits
 
 import io.vexelabs.bitbuilder.llvm.internal.contracts.ContainsReference
-import io.vexelabs.bitbuilder.raii.resourceScope
-import io.vexelabs.bitbuilder.raii.toResource
+import io.vexelabs.bitbuilder.internal.resourceScope
+import io.vexelabs.bitbuilder.internal.toResource
 import org.bytedeco.javacpp.IntPointer
 import org.bytedeco.llvm.LLVM.LLVMValueRef
 import org.bytedeco.llvm.global.LLVM
