+++ src/test/kotlin/io/vexelabs/bitbuilder/llvm/unit/ir/values/GlobalAliasTest.kt
@@ -4,7 +4,7 @@ import io.vexelabs.bitbuilder.llvm.ir.Module
 import io.vexelabs.bitbuilder.llvm.ir.types.IntType
 import io.vexelabs.bitbuilder.llvm.ir.values.constants.ConstantInt
 import io.vexelabs.bitbuilder.llvm.setup
-import io.vexelabs.bitbuilder.rtti.cast
+import io.vexelabs.bitbuilder.internal.cast
 import org.spekframework.spek2.Spek
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
